# Improving IndAffine, LeastSquares, Quadratic