--- a/docs/src/functions.md
+++ b/docs/src/functions.md
@@ -72,6 +72,5 @@ Linear
 LogBarrier
 Maximum
 Quadratic
-QuadraticIterative
 SumPositive
 ```