… src/functions/indAffineIterative.jl
… src/functions/leastSquaresIterative.jl
… test/test_quadratic.jl