In light of the recent discussions, I took a look to `IndAffine` to find that this type should probably be improved.

I’ve updated the **dense** case to perform as many operations as possible in-place. This resulted in performance improvements, see following comments.

@mfalt: do you think it’s possible to do similar things in the **sparse** case? I understand that using this low-level linear algebra routines is not the easiest thing, especially for sparse matrices/factorizations. Documentation is sometimes lacking?

**EDIT**

In light of the similar implementation details involved in handling these types, this PR now contains a restructuring of the `IndAffine`, `LeastSquares` and `Quadratic` types, together with a tentative improvement in their performance (by e.g. making as many operations as possible in-place).

**TODO**

- [x] Make sure to have as much as possible done in-place in `IndAffine`, `LeastSquares` and `Quadratic`.