@lostella 2017-07-21
When running the following

```julia
using ProximalOperators
using BenchmarkTools
srand(0)
# define nonempty affine subspace (A, b)
m, n = 50, 200
A = randn(m, n)
x0 = randn(n)
b = A*x0
# construct indicator function
f = IndAffine(A, b)
# time projections
x = randn(n)
y = similar(x)
@benchmark prox!(y, f, x)
```

**Previous results** (6d98bd8)

```
BenchmarkTools.Trial:
  memory estimate:  25.34 KiB
  allocs estimate:  11
  --------------
  minimum time:     21.464 μs (0.00% GC)
  median time:      22.799 μs (0.00% GC)
  mean time:        26.144 μs (5.31% GC)
  maximum time:     1.872 ms (93.83% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

**New results** (f94f86c, this PR)

```
BenchmarkTools.Trial:
  memory estimate:  48 bytes
  allocs estimate:  3
  --------------
  minimum time:     12.717 μs (0.00% GC)
  median time:      13.409 μs (0.00% GC)
  mean time:        14.195 μs (0.00% GC)
  maximum time:     282.460 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```