@mfalt 2017-07-21
Great work by the way, nice improvements. I think there is room for improvement in the sparse case, especially `RES = SparseArrays.CHOLMOD.Dense(f.A*x - f.b)` and `y .= x .- f.A'RRres`. Some places that look like allocations are probably not however, I think Dense() will just wrap the data for example. 

We might want to consider having different concrete types for the sparse and dense case (since they may require different storage). I had a long discussion with @mi-volodin yesterday, it may be beneficial to have seperate concrete types for the dense/sparse and direct/iterative versions of IndAffine. I have implementation for iterative IndAffine im my package, and I think @mi-volodin will propose a couple of different options for the direct and dense case.

PS: my implementation of iterative IndAffine is actually implementation of the more general graph problem, but with a linear term: `f([x;z]) = q'x+i(Ax-βz==b)`, where β is 1 or -1.