@lostella 2017-07-21
I agree, this will also lighten the type parameters significantly. I guess something like

```julia
abstract type IndAffine <: ProximableFunction end
struct IndAffineDense <: IndAffine [...] end
struct IndAffineSparse <: IndAffine [...] end
struct IndAffineIterative <: IndAffine [...] end
```

And similarly for `Quadratic` and `LeastSquares`, for example. This has also the advantage that all predicates `is_smooth`, `is_quadratic` etc can be defined over the `abstract type` once and for all.