@mi-volodin 2017-07-21
This is what we have already suggested in my `IndGraph` project.
But in a manner
```
struct IndGraphFat <: IndGraph [...] end
struct IndGraphSkinny <: IndGraph [...] end
```
Because efficient methods for `(m<n)` and `(m>n)` are different.