@lostella 2017-07-21
Ok, great! As for the suggestion of @mfalt regarding the `@blas!` macro: I'll give it a try, but let's see if it's possible to do this without increasing package dependencies. I'm already quite satisfied by the improvement in time/allocations (especially allocations).