@mfalt 2017-07-21
The macro is just a short and easy command to call the built in functions, you can use macroexpand to see what is it's actually doing to avoid any dependencies, it's just calling axpy! I'm guessing.