@lostella 2017-07-24
I have updated the PR description above: the edit mainly refers to the latest commit 3d0cd80. This organization of `IndAffine`, `LeastSquares`, `Quadratic` makes very much sense to me: for the moment the only necessary "type-splits" I saw were between the "direct" and "iterative" case. A further split could make two separate types for the "direct-dense" and "direct-sparse", but I would avoid this and rely on dispatch, unless we see it's necessary.