@lostella 2017-08-21
I couldn't do anything more than increasing the tolerance for evaluating `f::IndAffine` in order to overcome the issue highlighted by @mi-volodin in #40.

@mfalt: it looks like in fact the `shift` option is **not** availabel for dense Cholesky factorizations. This is really unfortunate.

If nobody thinks there are big issues here, I would consider merging this PR. I decided to remove other goals from this PR to then open appropriate issues: smaller PRs are faster to merge, and it's easier to reason on smaller individual problems.