@mfalt 2017-08-21
I think it's ready to merge.

On Mon, Aug 21, 2017, 21:14 Lorenzo Stella <notifications@github.com> wrote:

> I couldn't do anything more than increasing the tolerance for evaluating
> f::IndAffine in order to overcome the issue highlighted by @mi-volodin
> <https://github.com/mi-volodin> in #40
> <https://github.com/kul-forbes/ProximalOperators.jl/pull/40>.
>
> @mfalt <https://github.com/mfalt>: it looks like in fact the shift option
> is *not* availabel for dense Cholesky factorizations. This is really
> unfortunate.
>
> If nobody thinks there are big issues here, I would consider merging this
> PR. I decided to remove other goals from this PR to then open appropriate
> issues: smaller PRs are faster to merge, and it's easier to reason on
> smaller individual problems.
>
> —
> You are receiving this because you were mentioned.
>
> Reply to this email directly, view it on GitHub
> <https://github.com/kul-forbes/ProximalOperators.jl/pull/41#issuecomment-323827417>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ADFlSXg0Ez8XTxBMYecgMP4f9SwDzYiHks5sadcngaJpZM4OfO6S>
> .
>